CI builds: Set max job count for cargo builds to 8
This is an attempt to ease the load of parallel builds on our CI
servers. With the shift from mostly node to rust builds, my hypothesis
is that now the CI server is often overloaded, leading to some
issues in timing-based tests (yes we have those), and partial
malfunctioning of Azure Pipelines agents.

This is a test, whether we can keep on having the same amount of agents.