Accessory Service(?): Standalone binary for reading and publishing card serial numbers (#5642)
Part of the Webview runtime on Android is the routing through of read card serial numbers from attached readers – either internal ones or Baltech. For the internal readers we're using the Android provided APIs, for Baltech a custom protocol.

To match that functionality on Windows (and also on Linux), we have to obviously use different APIs.

First option is to use the [PCSC](https://en.wikipedia.org/wiki/PC/SC) protocol, which uses the WinSmartcard API. On unixy systems, there's a free implementation: PC/SC-Lite. Problem is, that reading the card's serial number is highly specific to the card, and there's no standard reading it. Fortunately, most of the common chips offer a common command to read the card's serial number, hiding the difference between different card types. But Baltech doesn't support that. This is why this PR uses two mechanisms to read the card's serial number:
1) Use the Baltech SDK if a Baltech reader is attached (a dynamic library is needed),
2) Use PCSC for every other enabled PCSC reader (this has been successfully tested with the ET51-internal reader, and a common ACR122u USB reader).

Hotplugging is supported.

All card readings will be published via the `actyxos_sdk`:
```
➜  bin git:(ow/pcsc-bin) ✗ ./query_store --live '[{"semantics":"NO_IMAGINATION"}]'
using offsets {"b1QGPmrjyQ3":2}
{"lamport":321,"stream":{"semantics":"NO_IMAGINATION","name":"NO_IMAGINATION","source":"b1QGPmrjyQ3"},"timestamp":"2020-11-27T12:43:01Z.91967","offset":0,"payload":{"cardType":null,"reader":{"friendlyName":"ACS ACR122U PICC Interface 00 00"},"uid":[70,40,196,90]}}
{"lamport":322,"stream":{"semantics":"NO_IMAGINATION","name":"NO_IMAGINATION","source":"b1QGPmrjyQ3"},"timestamp":"2020-11-27T12:43:10Z.04819","offset":1,"payload":{"cardType":null,"reader":{"friendlyName":"Unknown"},"uid":[70,40,196,90]}}
{"lamport":323,"stream":{"semantics":"NO_IMAGINATION","name":"NO_IMAGINATION","source":"b1QGPmrjyQ3"},"timestamp":"2020-11-27T12:43:32Z.61632","offset":2,"payload":{"cardType":null,"reader":{"friendlyName":"ACS ACR122U PICC Interface 00 00"},"uid":[70,40,196,90]}}
switching
{"lamport":329,"stream":{"semantics":"NO_IMAGINATION","name":"NO_IMAGINATION","source":"b1QGPmrjyQ3"},"timestamp":"2020-11-27T12:45:52Z.71551","offset":3,"payload":{"cardType":null,"reader":{"friendlyName":"ACS ACR122U PICC Interface 00 00"},"uid":[70,40,196,90]}}
```

Confirmed working on both Windows (10) and Linux (with `pcscd` running).