Change the sourceid backwards compatibility approach.
Basically, every SourceId can now be converted to a NodeId, and some NodeIds
can be converted back to SourceIds.

A NodeId is assumed to be created from a SourceId when the last 16 bytes
are 0. The probability of this happening accidentally is astronomically low,
and in any case it does not matter if you treat NodeIds as opaque values.