rust/actyx/swarm: relax loading of known streams during startup.
On startup, known `StreamId`s are loaded from the attached sqlite index
store. Those are converted into an `StreamAlias`, and those are tried to
resolve to a cid with the attached IPFS daemon. If a stream was loaded,
but had not yet published any events to it, the StreamAlias might not
have been added. This should only happen during some interruptions
during the first startups if no events have been published so far.

Closes #6260