actyx/swarm: implement hard-coded ephemeral event streams (#6402)
This adds support for configuring a `RetainConfig` for individual
`StreamNr`s. Supported configuration modes are `MaxEvents`, which
retains the last n events of said stream, and `MaxAge`, which retains
the events that have been emitted within the `now-retain_duration`, and
MaxSize, which retains events up to the specified size.

Further, this is used to have static configs for the following streams:
* discovery (stream nr 1)
* metrics (stream nr 2)

Currently both hard-coded to retain the last 1000 events.