Switch to ReentrantSafeMutex for the main lock
Just in case we have a situation where we do an "upcall" from inside holding
the lock that causes some code to be executed that tries to take the lock again.

With a straight parking_lot mutex that would lead to a deadlock. With this,
it will lead to a panic, which should have the stack trace that you need to
figure out what is going on.