releases-new: refactorings
This refactors the original proposal from @o1iver a little bit. The
basic idea stays the same though.
Notable changes:
* versions file format is changed, such that there is a total order.
This decreases readability a little bit, but has the advantage that the
whole file can be regenerated by `releases-new` on the fly. This should
be prove to be a more stable on-disk format.
* Adds the `release` subcommand, which will calculate the set of version
changes based on the versions persisted in the given input file. If
there are new versions, the following happens:
  1) Create new branch `release/<HEAD ID>`;
  2) Commit changes to input file; the changelog will be placed into the
  commit's message;
  3) Push new branch to `origin`.