cosmos-release: properly handle multi-arch manifests (#7149)
The former approach of using `docker image {tag,push}` to release the
built multi-arch images did not work as intended, as they operated only
on one specific image of the source multi-arch manifest (namely the one
for the OS `cosmos-release` was running on). The only way short of
re-building the multi-arch image using `docker buildx build` is to
manually construct the manifest.
Another fix included in this commit is done to the existence check of
published docker manifests. So far, only the existence of a specific tag
was checked, which makes sense for explicit versions, but not for
re-used pointers like `latest`. What should actually be checked is
whether this tag points to the right manifest.

Closes #7145