Extend gossip protocol (#7335)
fix(actyx): Fill the `toReplicate` map returned by the `/offsets` endpoint as early as possible.

This commit extends the current gossip protocol: Offsets and lamport timestamps for published trees are added to the RootUpdate and RootMap structs. The encoding of said structs is changed to be extensible, such that additional members can be added in the future without needing to special handle decoding (like introduced here for the Actyx <=2.3.0).
The wire format is backwards compatible.

**Motivation**
The /offsets endpoint returns a toReplicate map, which lists the number of events pending replication per stream. This offset behind this API is driven by the ingested RootUpdates or RootMaps messages. However, the referenced tree header needs to be resolved to get the offset / event count. This means that toReplicate is mostly empty (apart from cases where the rest of the tree is very big or the network very slow). A scenario, where gosipped (slow path or root maps) messages reach the node, but resolving bits via bitswap was thus not observable.
Now one might argue that whenever gossip works, bitswap also works. One the one hand, this is not as resilient as our system can be, on the other hand, this will need to change anyway once we have selective replication.

This feature is needed for #6391.