Add `waitForSync` to TS SDK
This adds the `/api/v2/node/info` (authenticated) endpoint to Actyx,
which returns the following response:
```
{
  "connectedNodes": 0,
  "uptime": {
    "secs": 39,
    "nanos": 900657761
  },
  "version": "unknown-75493e4637bff572d9a05ac0946f95c3d275b511_dirty-linux-x86_64-debug"
}
```
This, in conjunction with the /api/v2/events/offsets response is used to
provide a waitForSync function in the TS SDK offered as a method on the
`Actyx` object: Wait for the connected node to be in sync with the
swarm. This is on a best-effort basis and waits at most 30 seconds after
the point in time the node has been started. This can be used in order
to reduce stale state inside the application when started together with
an Actyx node.

feat(ts-sdk): add waitForSwarmSync method.