Rollback the publishing method
The manifest create & push mechanism is a way of
"promoting images to production", by creating a manifest and pushing it under the
the name of another image, Docker will essentially just re-use the underlying images
to create a new one, under a new namespace. This is *crucial* for us because we
check if Docker releases are missing using the manifests.